fix: move env check file
